+++ polkadot/node/subsystem-util/src/lib.rs
@@ -49,6 +49,7 @@ use polkadot_primitives::{
 	ValidatorSignature,
 };
 pub use rand;
+use runtime::get_disabled_validators_with_fallback;
 use sp_application_crypto::AppCrypto;
 use sp_core::ByteArray;
 use sp_keystore::{Error as KeystoreError, KeystorePtr};
@@ -57,7 +58,6 @@ use std::{
 	time::Duration,
 };
 use thiserror::Error;
-use vstaging::get_disabled_validators_with_fallback;
 
 pub use determine_new_blocks::determine_new_blocks;
 pub use metered;
