+++ polkadot/runtime/parachains/src/paras/tests.rs
@@ -16,7 +16,7 @@
 
 use super::*;
 use frame_support::{assert_err, assert_ok, assert_storage_noop};
-use polkadot_primitives::{vstaging::SchedulerParams, BlockNumber, PARACHAIN_KEY_TYPE_ID};
+use polkadot_primitives::{BlockNumber, SchedulerParams, PARACHAIN_KEY_TYPE_ID};
 use polkadot_primitives_test_helpers::{dummy_head_data, dummy_validation_code, validator_pubkeys};
 use sc_keystore::LocalKeystore;
 use sp_keyring::Sr25519Keyring;
